# swagger.yaml
openapi: 3.0.3
info:
  title: iSchool
  description: >-
    # A simple Student Management System
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com, whatapp: +2340935095173 }
servers:
  - url: https://jigipay.onrender.com/api/v1
    description: This is the original api URL
    schemes: [https, http]
  - url: http://localhost:8080/api/v1
    description: THis is my local Server
paths:
...
  /admins:
    get:
      tags:
        - 👨‍💼 Admin

      summary: Get admins
      description: >-
        The API endpoint retrieves a list of admins. When accessing this
        endpoint, you will receive a response containing list of app users.

        It provides a convenient way to obtain admins information for
        various purposes within your application.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '10'
      responses:
        '200':
          description: Get admins
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '10757'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:13 GMT
            ETag:
              schema:
                type: string
                example: W/"2a05-xjkqHU4sNAeFPZX9uGNel7XPn8E"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      currentPageItems:
                        type: number
                        example: 10
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            email:
                              type: string
                              example: joseph.evans@example.com
                            gender:
                              type: string
                              example: male
                            _id:
                              type: number
                              example: 1
                            firstName:
                              type: string
                              example: Joseph
                            lastName:
                              type: string
                              example: Evans
                            phone:
                              type: string
                              example: (612)-327-2806
                            avatar:
                              type: object
                              properties:
                                public_id:
                                  type: string
                                  example: 98789j-98078998678u
                                url:
                                  type: string
                                  example: >-
                                    https://randomuser.me/api/portraits/med/men/65.jpg
                        example:
                          email: joseph.evans@example.com
                          gender: male
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: User
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                      limit:
                        type: number
                        example: 10
                      nextPage:
                        type: boolean
                        example: true
                      page:
                        type: number
                        example: 1
                      previousPage:
                        type: boolean
                        example: false
                      totalItems:
                        type: number
                        example: 500
                      totalPages:
                        type: number
                        example: 50
                  message:
                    type: string
                    example: Admins fetched successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Get admins:
                  value:
                    data:
                      currentPageItems: 10
                      data:
                        - email: joseph.evans@example.com
                          gender: male
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: ADMIN
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                        - email: john.evans@example.com
                          gender: female
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: ADMIN
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                      limit: 10
                      nextPage: true
                      page: 1
                      previousPage: false
                      totalItems: 500
                      totalPages: 50
                    message: Admins fetched successfully
                    success: true
  /admins/users:
    get:
      tags:
        - 👨‍💼 Admin

      summary: Get users
      description: >-
        The API endpoint retrieves a list of users. When accessing this
        endpoint, you will receive a response containing list of app users.

        It provides a convenient way to obtain user information for
        various purposes within your application.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '10'
      responses:
        '200':
          description: Get users
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '10757'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:13 GMT
            ETag:
              schema:
                type: string
                example: W/"2a05-xjkqHU4sNAeFPZX9uGNel7XPn8E"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      currentPageItems:
                        type: number
                        example: 10
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            email:
                              type: string
                              example: joseph.evans@example.com
                            gender:
                              type: string
                              example: male
                            _id:
                              type: number
                              example: 1
                            firstName:
                              type: string
                              example: Joseph
                            lastName:
                              type: string
                              example: Evans
                            phone:
                              type: string
                              example: (612)-327-2806
                            avatar:
                              type: object
                              properties:
                                public_id:
                                  type: string
                                  example: 98789j-98078998678u
                                url:
                                  type: string
                                  example: >-
                                    https://randomuser.me/api/portraits/med/men/65.jpg
                        example:
                          - email: joseph.evans@example.com
                            gender: male
                            _id: 08909789789
                            firstName: Joseph
                            username: Joseph
                            lastName: Evans
                            role: User
                            phone: (612)-327-2806
                            avatar:
                              public_id: 097897986ij0i
                              url: >-
                                https://randomuser.me/api/portraits/med/men/65.jpg
                          - email: joseph.evans@example.com
                            gender: male
                            _id: 08909789789
                            firstName: Joseph
                            username: Joseph
                            lastName: Evans
                            role: User
                            phone: (612)-327-2806
                            avatar:
                              public_id: 097897986ij0i
                              url: >-
                                https://randomuser.me/api/portraits/med/men/65.jpg
                      limit:
                        type: number
                        example: 10
                      nextPage:
                        type: boolean
                        example: true
                      page:
                        type: number
                        example: 1
                      previousPage:
                        type: boolean
                        example: false
                      totalItems:
                        type: number
                        example: 500
                      totalPages:
                        type: number
                        example: 50
                  message:
                    type: string
                    example: Users fetched successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Get users:
                  value:
                    data:
                      currentPageItems: 10
                      data:
                        - email: joseph.evans@example.com
                          gender: male
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: User
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                        - email: john.evans@example.com
                          gender: female
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: User
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                      limit: 10
                      nextPage: true
                      page: 1
                      previousPage: false
                      totalItems: 500
                      totalPages: 50
                    message: user fetched successfully
                    success: true
...
  /admins/transactions:
    get:
      tags:
        - 👨‍💼 Admin
      summary: Get transactions
      description: >-
        The API endpoint retrieves a list of transactions. When accessing this
        endpoint, you will receive a response containing list of users transactions.

        It provides a convenient way to obtain user transaction information for
        various purposes within your application.
      operationId: getTransactions
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '10'
      responses:
        '200':
          description: Get transactions
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '10757'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:13 GMT
            ETag:
              schema:
                type: string
                example: W/"2a05-xjkqHU4sNAeFPZX9uGNel7XPn8E"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      currentPageItems:
                        type: number
                        example: 10
                        # edit data
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: 649e795264e7dba29b726936
                            userId:
                              type: string
                              example: 649e795264e7dba29b726936
                            receiverId:
                              type: string
                              example: 649e795264e7dba29b726336
                            bank:
                              type: string
                              example: 'name of bank'
                            account:
                              type: number
                              example: 008080909
                            accountName:
                              type: string
                              example: jamilu bello
                            description:
                              type: string
                              example: 'a simple transaction'
                            naration:
                              type: string
                              example: 'a simple naration'
                            balance:
                              type: number
                              example: 1222
                            amount:
                              type: number
                              example: 1222
                            status:
                              type: string
                              example: 'SUCCESS'
                            type:
                              type: string
                              example: 'CREDIT'
                            updatedAt:
                              type: string
                              example: '2023-06-30T06:44:29.831Z'

                        example:
                          email: joseph.evans@example.com
                          gender: male
                          _id: 08909789789
                          firstName: Joseph
                          username: Joseph
                          lastName: Evans
                          role: User
                          phone: (612)-327-2806
                          avatar:
                            public_id: 097897986ij0i
                            url: >-
                              https://randomuser.me/api/portraits/med/men/65.jpg
                      limit:
                        type: number
                        example: 10
                      nextPage:
                        type: boolean
                        example: true
                      page:
                        type: number
                        example: 1
                      previousPage:
                        type: boolean
                        example: false
                      totalItems:
                        type: number
                        example: 500
                      totalPages:
                        type: number
                        example: 50
                  message:
                    type: string
                    example: Transactions fetched successfully
                  success:
                    type: boolean
                    example: true
              examples:
                # edit title
                Get transactions:
                  value:
                    data:
                      currentPageItems: 10
                      # edit data
                      data:
                        - userId: 649e795264e7dba29b96936
                          _id: 649e795264e7dba29b726036
                          recieverId: 649e795264e7dba29b96976
                          bank: jigipay
                          amount: 900
                          type: DEBIT
                          status: SUCCESS
                          descrption: a simple description
                          naration: a simple naration
                          updatedAt: '2023-06-30T06:44:29.831Z'
                        - _id: 649e795264e7dba29b726036
                          userId: 649e795264e7dba29b96936
                          accountName: musa kano
                          bank: first bank
                          amount: 900
                          balance: 120
                          type: DEBIT
                          status: SUCCESS
                          descrption: a simple description
                          naration: a simple naration
                          updatedAt: '2023-06-30T06:44:29.831Z'
                      limit: 10
                      nextPage: true
                      page: 1
                      previousPage: false
                      totalItems: 500
                      totalPages: 50
                    message: user fetched successfully
                    success: true
...